moved getInitials function to Utils
