+++ src/components/Person/Person.jsx
@@ -1,17 +1,10 @@
 import React from 'react';
 import {NavLink} from 'react-router-dom';
+import {getInitials} from '../../Utils/getInitials';
 
 import './person.css';
 
 export const Person = ({name, birth_year, gender, url}) => {
-  const getInitials = name => {
-    const words = name.split(/[\s-]+/);
-    let initials = '';
-    for (let i = 0; i < words.length; i++) {
-      initials += words[i].charAt(0).toUpperCase();
-    }
-    return initials;
-  };
   const initials = getInitials(name);
 
   const personId = url.match(/\/(\d+)/)[1];
